refactor(ui): Move UI type definition/redefinition file
The file used for defining some types missing from the API (advisors,
package managers, report formats) was located deep in the file 
hierarchy. As the usage of this file is not really localized anymore, 
move the file to the `ui/src/lib` folder for easier access and simpler
imports.
